refactor(features): move main function to root src directory
